requirements.txt
Faker
Flask
Flask-SQLAlchemy
importlib-metadata
importlib-resources
sqlalchemy-serializer
sqlalchemy
ipdb
pytest
flask-migrate
flask-restful
flask-swagger-ui
gunicorn
python-dotenv
flask-cors